+++ lib/src/context/mod.rs
@@ -1,2 +1,3 @@
 pub mod session;
 pub mod session_manager;
+pub mod variable;