+++ lib/src/response/handler/mod.rs
@@ -3,3 +3,4 @@ pub mod dataframe;
 pub mod processor;
 pub mod session;
 pub mod sys_info;
+pub mod variable;